requirements.txt
pytz==2020.1
discord.py==1.5.0
requests==2.24.0